working prototype of the project folder creation system

+++ src/projectSystem.py
@@ -4,22 +4,32 @@ import os
 rootProjectDir = "QSYS-Projects"
 
 class Project:
-    #initilization of the project with the name and the path
-    def __init__(self, name, path):
+    #initilization of the project with the name and the path located in a folder named QSYS-Projects
+    def __init__(self, name):
         self.name = name
-        self.path = f'{path}'
+        # self.path = f'./{rootProjectDir}/{name}'
+        self.path = f'{name}'
         self.files = []
         self.folders = []
 
     # create a project with the name and the path then create the directory and the folders and files
     def createProject(self):
-        os.mkdir(self.path)
-        os.chdir(self.path)
-        for folder in self.folders:
-            os.mkdir(folder)
-        for file in self.files:
-            open(file, 'w')
-        self.createProjectReadme()
+        if os.path.exists(rootProjectDir):
+            os.chdir(rootProjectDir)
+        else:
+            os.mkdir(rootProjectDir)
+            os.chdir(rootProjectDir)
+        if os.path.exists(self.path):
+            print(f'Project {self.name} already exists')
+        else:
+            os.mkdir(self.path)
+            os.chdir(self.path)
+            self.createProjectReadme()
+            for folder in self.folders:
+                os.mkdir(folder)
+            for file in self.files:
+                open(file, 'w')
+            self.createProjectReadme()
 
     def createProjectReadme(self):
         with open('README.md', 'w') as readme:
@@ -29,7 +39,7 @@ class Project:
             readme.write(f'Folders: {self.folders}\n')
 
 if __name__ == "__main__":
-    project = Project("ProjectName-Test", "ProjectName-Test-Milwaukee-WI")
+    project = Project(input('Project Name: '))
     project.folders = ["resources", ".backup-rev"]
     project.files = ["testproject.qsys", "Readme.md"]
     project.createProject()
